Algorithms and Deep Learning Models/SkillSync/requirements.txt
Flask
requests
google-generativeai
pandas
dotenv
